src/assimilator_agent/requirements.txt
llama-index
llama-index-embeddings-ollama
llama-index-llms-openai
llama-index-llms-ollama
llama-index-readers-wikipedia
llama-index-agent-llm-compiler
llama-index-readers-smart-pdf-loader
wikipedia
llmsherpa
python-dotenv
fastapi
python-multipart
uvicorn